Extract symbolic extraction routines into symbolic_extraction
