+++ drake/solvers/BUILD
@@ -44,6 +44,21 @@ drake_cc_library(
 )
 
 drake_cc_library(
+    name = "symbolic_extraction",
+    srcs = ["symbolic_extraction.cc"],
+    hdrs = ["symbolic_extraction.h"],
+    deps = [
+        ":decision_variable",
+        "//drake/common",
+        "//drake/common:autodiff",
+        "//drake/common:monomial",
+        "//drake/common:number_traits",
+        "//drake/common:polynomial",
+        "//drake/common:symbolic",
+    ],
+)
+
+drake_cc_library(
     name = "create_cost",
     srcs = [],
     hdrs = ["create_cost.h"],
@@ -122,6 +137,7 @@ drake_cc_library(
         ":mosek_solver",
         ":nlopt_solver",
         ":snopt_solver",
+        ":symbolic_extraction",
         "//drake/common",
         "//drake/common:autodiff",
         "//drake/common:monomial",
