+++ drake/solvers/mathematical_program.cc
@@ -25,6 +25,7 @@
 #include "drake/solvers/mosek_solver.h"
 #include "drake/solvers/nlopt_solver.h"
 #include "drake/solvers/snopt_solver.h"
+#include "drake/solvers/symbolic_extraction.h"
 
 namespace drake {
 namespace solvers {
@@ -50,6 +51,12 @@ using std::vector;
 using symbolic::Expression;
 using symbolic::Formula;
 using symbolic::Variable;
+
+using internal::DecomposeLinearExpression;
+using internal::DecomposeQuadraticExpressionWithMonomialToCoeffMap;
+using internal::ExtractAndAppendVariablesFromExpression;
+using internal::ExtractVariablesFromExpression;
+using internal::SymbolicError;
 
 namespace {
 
@@ -200,234 +207,6 @@ VectorXDecisionVariable MathematicalProgram::NewBinaryVariables(
 }
 
 namespace {
-
-class SymbolicError : public runtime_error {
- public:
-  SymbolicError(const Expression& e, const string& msg)
-      : runtime_error{make_string(e, msg)} {}
-  SymbolicError(const Expression& e, const double lb, const double ub,
-                const string& msg)
-      : runtime_error{make_string(e, lb, ub, msg)} {}
-
- private:
-  static string make_string(const Expression& e, const string& msg) {
-    ostringstream oss;
-    oss << "Constraint " << e << " is " << msg << ".";
-    return oss.str();
-  }
-  static string make_string(const Expression& e, const double lb,
-                            const double ub, const string& msg) {
-    ostringstream oss;
-    oss << "Constraint " << lb << " <= " << e << " <= " << ub << " is " << msg
-        << ".";
-    return oss.str();
-  }
-};
-
-// Given an expression `e`, extracts all variables inside `e`.
-// @param[in] e A symbolic expression.
-// @retval pair pair.first is the variables in `e`. pair.second is the mapping
-// from the variable ID to the index in pair.first, such that
-// pair.second[pair.first(i).get_id()] = i
-pair<VectorXDecisionVariable, unordered_map<Variable::Id, int>>
-ExtractVariablesFromExpression(const Expression& e) {
-  int var_count = 0;
-  const symbolic::Variables var_set = e.GetVariables();
-  VectorXDecisionVariable vars(var_set.size());
-  unordered_map<Variable::Id, int> map_var_to_index{};
-  map_var_to_index.reserve(var_set.size());
-  for (const Variable& var : var_set) {
-    map_var_to_index.emplace(var.get_id(), var_count);
-    vars(var_count++) = var;
-  }
-  return make_pair(vars, map_var_to_index);
-}
-
-// Given an expression `e`, extract all variables inside `e`, append these
-// variables to `vars` if they are not included in `vars` yet.
-// @param[in] e  A symbolic expression.
-// @param[in,out] vars  As an input, `vars` contain the variables before
-// extracting expression `e`. As an output, the variables in `e` that were not
-// included in `vars`, will be appended to the end of `vars`.
-// @param[in,out] map_var_to_index. map_var_to_index is of the same size as
-// `vars`, and map_var_to_index[vars(i).get_id()] = i. This invariance holds
-// for map_var_to_index both as the input and as the output.
-void ExtractAndAppendVariablesFromExpression(
-    const Expression& e, VectorXDecisionVariable* vars,
-    unordered_map<Variable::Id, int>* map_var_to_index) {
-  DRAKE_DEMAND(static_cast<int>(map_var_to_index->size()) == vars->size());
-  for (const Variable& var : e.GetVariables()) {
-    if (map_var_to_index->find(var.get_id()) == map_var_to_index->end()) {
-      map_var_to_index->emplace(var.get_id(), vars->size());
-      vars->conservativeResize(vars->size() + 1, Eigen::NoChange);
-      (*vars)(vars->size() - 1) = var;
-    }
-  }
-}
-
-/** Decomposes a linear combination @p e = c0 + c1 * v1 + ... cn * vn into
- *  the followings:
- *
- *     constant term      : c0
- *     coefficient vector : [c1, ..., cn]
- *     variable vector    : [v1, ..., vn]
- *
- *  Then, it extracts the coefficient and the constant term.
- *  A map from variable ID to int, @p map_var_to_index, is used to decide a
- *  variable's index in a linear combination.
- *
- *  \pre{1. @c coeffs is a row vector of double, whose length matches with the
- *          size of @c map_var_to_index.
- *       2. e.is_polynomial() is true.
- *       3. e is a linear expression.}
- * @tparam Derived An Eigen row vector type with Derived::Scalar == double.
- * @param[in] e The symbolic linear expression
- * @param[in] map_var_to_index A mapping from variable ID to variable index,
- * such that map_var_to_index[vi.get_ID()] = i.
- * @param[out] coeffs A row vector. coeffs(i) = ci.
- * @param[out] constant_term c0 in the equation above.
- * @return num_variable. Number of variables in the expression. 2 * x(0) + 3
- * has 1 variable, 2 * x(0) + 3 * x(1) - 2 * x(0) has 1 variable.
- */
-template <typename Derived>
-typename enable_if<is_same<typename Derived::Scalar, double>::value, int>::type
-DecomposeLinearExpression(
-    const Expression& e,
-    const unordered_map<Variable::Id, int>& map_var_to_index,
-    const Eigen::MatrixBase<Derived>& coeffs, double* constant_term) {
-  DRAKE_DEMAND(coeffs.rows() == 1);
-  DRAKE_DEMAND(coeffs.cols() == static_cast<int>(map_var_to_index.size()));
-  if (!e.is_polynomial()) {
-    ostringstream oss;
-    oss << "Expression " << e << "is not a polynomial.\n";
-    throw runtime_error(oss.str());
-  }
-  const symbolic::Variables& vars = e.GetVariables();
-  const auto& monomial_to_coeff_map =
-      symbolic::DecomposePolynomialIntoMonomial(e, vars);
-  int num_variable = 0;
-  for (const auto& p : monomial_to_coeff_map) {
-    const auto& p_monomial = p.first;
-    DRAKE_ASSERT(is_constant(p.second));
-    const double p_coeff = symbolic::get_constant_value(p.second);
-    if (p_monomial.total_degree() > 1) {
-      ostringstream oss;
-      oss << "Expression " << e << " is non-linear.";
-      throw runtime_error(oss.str());
-    } else if (p_monomial.total_degree() == 1) {
-      // Linear coefficient.
-      const auto& p_monomial_powers = p_monomial.get_powers();
-      DRAKE_DEMAND(p_monomial_powers.size() == 1);
-      const Variable::Id var_id = p_monomial_powers.begin()->first;
-      const_cast<Eigen::MatrixBase<Derived>&>(coeffs)(
-          map_var_to_index.at(var_id)) = p_coeff;
-      if (p_coeff != 0) {
-        ++num_variable;
-      }
-    } else {
-      // Constant term.
-      *constant_term = p_coeff;
-    }
-  }
-  return num_variable;
-}
-
-/**
- * Given a vector of linear expressions v, decompose it to
- * \f$ v = A vars + b \f$
- * @param[in] v A vector of linear expressions
- * @param[out] A The matrix containing the linear coefficients.
- * @param[out] b The vector containing all the constant terms.
- * @param[out] vars All variables.
- */
-void DecomposeLinearExpression(const Eigen::Ref<const VectorX<Expression>>& v,
-                               Eigen::MatrixXd* A, Eigen::VectorXd* b,
-                               VectorXDecisionVariable* vars) {
-  // 0. Setup map_var_to_index and var_vec.
-  unordered_map<Variable::Id, int> map_var_to_index;
-  for (int i = 0; i < v.size(); ++i) {
-    ExtractAndAppendVariablesFromExpression(v(i), vars, &map_var_to_index);
-  }
-
-  // 2. Construct decompose v as
-  // v = A * vars + b
-  *A = Eigen::MatrixXd::Zero(v.rows(), vars->rows());
-  *b = Eigen::VectorXd::Zero(v.rows());
-  for (int i{0}; i < v.size(); ++i) {
-    const Expression& e_i{v(i)};
-    DecomposeLinearExpression(e_i, map_var_to_index, A->row(i), b->data() + i);
-  }
-}
-
-/**
- * Given a quadratic expression `e` represented by its monomial to coefficient
- * map, decompose it into the form
- * e = 0.5 * x' * Q * x + b' * x + c
- * @param[in] monomial_to_coeff_map. Map the monomial to the coefficient, this
- * is the result of calling DecomposePolynomialIntoMonomial(e).
- * @param[in] map_var_to_index maps variables in
- * monomial_to_coeff_map.GetVariables() to the index in the vector `x`.
- * @param[in] num_variables The number of variables in the expression.
- * @param Q[out] The Hessian of the quadratic expression. @pre The size of Q
- * should be `num_variables * num_variables`.
- * @param b[out] The linear term of the quadratic expression. @pre The size of
- * `b` should be `num_variables * 1`.
- * @param c[out] The constant term of the quadratic expression.
- */
-void DecomposeQuadraticExpressionWithMonomialToCoeffMap(
-    const symbolic::MonomialToCoefficientMap& monomial_to_coeff_map,
-    const unordered_map<Variable::Id, int>& map_var_to_index, int num_variables,
-    Eigen::MatrixXd* Q, Eigen::VectorXd* b, double* c) {
-  DRAKE_DEMAND(Q->rows() == num_variables);
-  DRAKE_DEMAND(Q->cols() == num_variables);
-  DRAKE_DEMAND(b->rows() == num_variables);
-  Q->setZero();
-  b->setZero();
-  *c = 0;
-  for (const auto& p : monomial_to_coeff_map) {
-    DRAKE_ASSERT(is_constant(p.second));
-    DRAKE_DEMAND(!is_zero(p.second));
-    const double coefficient = get_constant_value(p.second);
-    const symbolic::Monomial& p_monomial = p.first;
-    if (p_monomial.total_degree() > 2) {
-      ostringstream oss;
-      oss << p.first
-          << " has order higher than 2, cannot be handled by "
-             "DecomposeQuadraticExpressionWithMonomialToCoeffMap"
-          << endl;
-      throw runtime_error(oss.str());
-    }
-    const auto& monomial_powers = p_monomial.get_powers();
-    if (monomial_powers.size() == 2) {
-      // cross terms.
-      auto it = monomial_powers.begin();
-      const int x1_index = map_var_to_index.at(it->first);
-      DRAKE_DEMAND(it->second == 1);
-      ++it;
-      const int x2_index = map_var_to_index.at(it->first);
-      DRAKE_DEMAND(it->second == 1);
-      (*Q)(x1_index, x2_index) += coefficient;
-      (*Q)(x2_index, x1_index) = (*Q)(x1_index, x2_index);
-    } else if (monomial_powers.size() == 1) {
-      // Two cases
-      // 1. quadratic term a*x^2
-      // 2. linear term b*x
-      auto it = monomial_powers.begin();
-      DRAKE_DEMAND(it->second == 2 || it->second == 1);
-      const int x_index = map_var_to_index.at(it->first);
-      if (it->second == 2) {
-        // quadratic term a * x^2
-        (*Q)(x_index, x_index) += 2 * coefficient;
-      } else if (it->second == 1) {
-        // linear term b * x.
-        (*b)(x_index) += coefficient;
-      }
-    } else {
-      // constant term.
-      *c += coefficient;
-    }
-  }
-}
 
 template <typename To, typename From>
 Binding<To> BindingUpcast(const Binding<From>& binding) {
